minus fixed
